Rename eth_hacks to eth_utils
